requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlmodel>=0.0.14
pyjwt>=2.8.0
pyyaml>=6.0
bcrypt>=4.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
email-validator>=2.0.0
requests>=2.32.0
openai>=1.0.0
slowapi>=0.1.9
cryptography>=41.0.0
python-dotenv>=1.0.1
